feat(autoedit): Improve partial diff stability
